% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zum_aode.R
\name{aode_bayes}
\alias{aode_bayes}
\title{Creating aode object}
\usage{
aode_bayes(data, formula)
}
\arguments{
\item{data}{training data for model creation}

\item{formula}{variable dependency description}
}
\value{
aode object
}
\description{
Creating aode object
}
\examples{
data <- data.frame ( a1=c(1,1,2), a2=c(1,2,3), a3=c(1,1,2),c=c(0,0,1))
f <- c ~ .
x <- aode_bayes(data,f)
}
