% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zum_aode.R
\docType{package}
\name{bayesaode}
\alias{bayesaode}
\title{bayesaode: A package for classifying dataset using non-naive AODE Bayesian algorithm}
\description{
The package provides two functions: aode_bayes and predict for usage of
performing classification task using non-naive Bayesian algorithm:
}
\section{aode_bayes}{

aode_bayes function is used for creating model object for aode algorithm.
}

\section{predict}{

function is used to perform a classifaction task using aode object
}

